Refactor extension to inject accounts into page
This is a complete UI and architechtural overhaul, accounts are now listed in the settings/'cog'
dropdown menu on the twitter site itself. This is a vast improvement to user experience as it removes
the need to go 'out of window' (even if it was only a little) and is in keeping with the rest of the
twitter user experience.

I think this is how twitter would(/should) do it if they did it themselves, and I much prefer using it